refactor: import net-stubbing types from network-interception
Remove the net-stubbing type re-export shims (lib/types.ts,
external-types.ts, internal-types.ts) and update consumers to import
directly from @packages/network-interception. CLI sync-typedefs already
sources external-types from network-interception, so no CLI build change
was required.
