+++ knip.json
@@ -791,10 +791,6 @@
       "exports",
       "types"
     ],
-    "packages/net-stubbing/lib/external-types.ts": [
-      "exports",
-      "types"
-    ],
     "packages/network-interception/lib/types/external-types.ts": [
       "exports",
       "types"
