+++ packages/driver/src/cy/net-stubbing/events/network-error.ts
@@ -1,5 +1,5 @@
 import { get } from 'lodash'
-import type { CyHttpMessages } from '@packages/net-stubbing/lib/types'
+import type { CyHttpMessages } from '@packages/network-interception/lib/types'
 import $errUtils from '../../../cypress/error_utils'
 import type { HandlerFn } from '.'
