+++ packages/driver/src/cy/net-stubbing/events/utils.ts
@@ -1,5 +1,5 @@
 import { find } from 'lodash'
-import type { CyHttpMessages } from '@packages/net-stubbing/lib/types'
+import type { CyHttpMessages } from '@packages/network-interception/lib/types'
 
 function hasJsonContentType (headers: { [k: string]: string | string[] }) {
   const contentType = find(headers, (v, k) => /^content-type$/i.test(k))
