--- packages/net-stubbing/lib/external-types.ts
@@ -1,1 +1,0 @@
-export * from '@packages/network-interception/lib/types/external-types'
