--- packages/net-stubbing/lib/internal-types.ts
@@ -1,1 +1,0 @@
-export * from '@packages/network-interception/lib/types/internal-types'
