+++ packages/net-stubbing/lib/server/index.ts
@@ -12,8 +12,6 @@ export { handleInterceptResponse } from './handle-intercept-response'
 
 export { NetStubbingState, ResourceType } from './types'
 
-export { getRoutesForRequest } from '@packages/network-interception'
-
 import { state } from './state'
 
 export { state as netStubbingState }
