+++ packages/net-stubbing/lib/server/intercepted-request.ts
@@ -8,13 +8,14 @@ import type {
 import type {
   NetEvent,
   Subscription,
-} from '../types'
-import type { BackendRoute, NetStubbingState } from './types'
+  BackendStaticResponse,
+} from '@packages/network-interception/lib/types'
+import type { BackendRoute } from '@packages/network-interception'
+import type { NetStubbingState } from './types'
 import { planSubscriptions } from '@packages/network-interception'
 import * as errors from '@packages/errors'
 import { emit, sendStaticResponse } from './util'
 import type { SocketBroadcaster } from '@packages/socket'
-import type { BackendStaticResponse } from '../internal-types'
 
 export class InterceptedRequest {
   id: string
