+++ packages/net-stubbing/lib/server/util.ts
@@ -5,20 +5,18 @@ import isHtml from 'is-html'
 import { IncomingMessage } from 'http'
 import type {
   BackendStaticResponse,
-} from '../types'
-
-export { getAllStringMatcherFields } from '@packages/network-interception'
+} from '@packages/network-interception/lib/types'
 
 import { Readable, PassThrough } from 'stream'
 import { Socket } from 'net'
-import type { GetFixtureFn } from './types'
+import type { GetFixtureFn } from '@packages/network-interception'
 import ThrottleStream from 'throttle'
 import type { CypressIncomingRequest } from '@packages/proxy'
 import type { InterceptedRequest } from './intercepted-request'
 import type { SocketBroadcaster } from '@packages/socket'
 import { caseInsensitiveGet, caseInsensitiveHas } from '../util'
 
-import type { CyHttpMessages } from '../external-types'
+import type { CyHttpMessages } from '@packages/network-interception/lib/types/external-types'
 import { getEncoding } from 'istextorbinary'
 
 const debug = Debug('cypress:net-stubbing:server:util')
