--- packages/net-stubbing/lib/types.ts
@@ -1,3 +1,0 @@
-export * from './external-types'
-
-export * from './internal-types'
